requirements.txt
asgiref==3.2.10
attrs==20.2.0
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.7.0
cryptography==3.2.1
cssselect2==0.4.1
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.2
django-allauth==0.43.0
django-filter==2.4.0
ecdsa==0.14.1
Fiona==1.8.17
geographiclib==1.50
geopandas==0.8.1
geopy==2.0.0
gmane==0.1.dev25
gunicorn==20.0.4
html5lib==1.1
httplib2==0.18.1
idna==2.10
lxml==4.6.1
munch==2.5.0
networkx==2.5
numpy==1.19.3
oauth2==1.9.0.post1
oauthlib==3.1.0
pandas==1.1.4
Pillow==8.0.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
Pyphen==0.10.0
pyproj==2.6.1.post1
python-dateutil==2.8.1
python-jose==3.2.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.141.0
Shapely==1.7.1
six==1.15.0
social==0.1.dev3
social-auth-app-django==4.0.0
social-auth-core==3.3.3
splinter==0.14.0
sqlparse==0.4.1
tinycss2==1.1.0
urllib3==1.25.11
WeasyPrint==52.1
webencodings==0.5.1
whitenoise==5.2.0
